Move QueryVectorBuilderAsyncAction into a dedicated file
