+++ server/src/main/java/org/elasticsearch/search/vectors/KnnVectorQueryBuilder.java
@@ -16,8 +16,6 @@ import org.apache.lucene.search.join.BitSetProducer;
 import org.apache.lucene.search.join.ToChildBlockJoinQuery;
 import org.apache.lucene.util.SetOnce;
 import org.elasticsearch.TransportVersion;
-import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.client.internal.Client;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.lucene.search.Queries;
@@ -31,7 +29,6 @@ import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.MatchNoneQueryBuilder;
 import org.elasticsearch.index.query.NestedQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
-import org.elasticsearch.index.query.QueryRewriteAsyncAction;
 import org.elasticsearch.index.query.QueryRewriteContext;
 import org.elasticsearch.index.query.SearchExecutionContext;
 import org.elasticsearch.index.query.ToChildBlockJoinQueryBuilder;
@@ -443,18 +440,7 @@ public class KnnVectorQueryBuilder extends AbstractQueryBuilder<KnnVectorQueryBu
         }
         if (queryVectorBuilder != null) {
             SetOnce<float[]> toSet = new SetOnce<>();
-            ctx.registerUniqueAsyncAction(new QueryVectorBuilderAsyncAction(queryVectorBuilder), v -> {
-                toSet.set(v);
-                if (v == null) {
-                    throw new IllegalArgumentException(
-                        format(
-                            "[%s] with name [%s] returned null query_vector",
-                            QUERY_VECTOR_BUILDER_FIELD.getPreferredName(),
-                            queryVectorBuilder.getWriteableName()
-                        )
-                    );
-                }
-            });
+            QueryVectorBuilderAsyncAction.registerAction(ctx, queryVectorBuilder, toSet);
             return new KnnVectorQueryBuilder(
                 fieldName,
                 queryVector,
@@ -678,28 +664,5 @@ public class KnnVectorQueryBuilder extends AbstractQueryBuilder<KnnVectorQueryBu
     public KnnVectorQueryBuilder setAutoPrefilteringEnabled(boolean isAutoPrefilteringEnabled) {
         this.isAutoPrefilteringEnabled = isAutoPrefilteringEnabled;
         return this;
-    }
-
-    private static final class QueryVectorBuilderAsyncAction extends QueryRewriteAsyncAction<float[], QueryVectorBuilderAsyncAction> {
-        private final QueryVectorBuilder queryVectorBuilder;
-
-        private QueryVectorBuilderAsyncAction(QueryVectorBuilder queryVectorBuilder) {
-            this.queryVectorBuilder = Objects.requireNonNull(queryVectorBuilder);
-        }
-
-        @Override
-        protected void execute(Client client, ActionListener<float[]> listener) {
-            queryVectorBuilder.buildVector(client, listener);
-        }
-
-        @Override
-        public int doHashCode() {
-            return Objects.hash(queryVectorBuilder);
-        }
-
-        @Override
-        public boolean doEquals(QueryVectorBuilderAsyncAction other) {
-            return Objects.equals(queryVectorBuilder, other.queryVectorBuilder);
-        }
     }
 }
